fix: repository 변경

+++ deposit-system/src/main/java/com/bank/deposit/repository/LedgerRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LedgerRepository extends JpaRepository<Ledger, Integer> {
+public interface LedgerRepository extends JpaRepository<Ledger, Long> {
 }
